serveur/requirements_client.txt
requests
websocket
dotenv